Fix file format identifier for input files in setup instructions
The instructions ask the learner to create a file "chr1-star-index.yaml"
that references an input file in the FASTA format, if they are to follow
the steps for generating the STAR index themselves.

In the YAML file however, the "format" field for that file referred to
something else (FASTQ format).  If the user followed the instructions
and submitted the workflow to Arvados, the run would fail validation.

Fixed by using the intended URI for the format (tested in Arvados 2.8.0
dev).

Arvados-DCO-1.1-Signed-off-by: Zoë Ma <zoe.ma@curii.com>